+++ packages/types/src/exception.ts
@@ -1,9 +1,11 @@
+import { Mechanism } from './mechanism';
 import { Stacktrace } from './stacktrace';
 
 /** JSDoc */
 export interface Exception {
   type?: string;
   value?: string;
+  mechanism?: Mechanism;
   module?: string;
   thread_id?: number;
   stacktrace?: Stacktrace;