+++ packages/opentracing/src/index.ts
@@ -1,0 +1,3 @@
+export { Span } from './span';
+export { Tracer } from './tracer';
+export { OpenTracingIntegration as Integration } from './opentracingintegration';