+++ packages/browser/test/integration/dedupe.js
@@ -1,1 +1,0 @@
-../../../integrations/build/dedupe.js