+++ dangerfile.ts
@@ -5,7 +5,7 @@ import { resolve } from 'path';
 import tslint from 'danger-plugin-tslint';
 import { prettyResults } from 'danger-plugin-tslint/dist/prettyResults';
 
-const packages = ['browser', 'core', 'hub', 'integrations', 'minimal', 'node', 'types', 'utils'];
+const packages = ['browser', 'core', 'hub', 'integrations', 'opentracing', 'minimal', 'node', 'types', 'utils'];
 
 export default async () => {
   if (!danger.github) {
@@ -48,6 +48,6 @@ export default async () => {
 
   schedule(async () => {
     const result = (await promisify(exec)('cd packages/browser; yarn size:check')).stdout;
-    message(`@sentry/browser bundle gzip'ed minified size: \n${result.split('\n')[2]}\n${result.split('\n')[3]}`);
+    message(`@sentry/browser bundle gzip'ed minified size: \n${result.split('\n')[4]}\n${result.split('\n')[5]}`);
   });
 };