+++ CHANGELOG.md
@@ -35,6 +35,8 @@ since we removed some methods from the public API and removed some classes from 
 - **breaking** [core] ref: Move `extraErrorData` integration to `@sentry/integrations` package
 - [core] feat: Add `maxValueLength` option to adjust max string length for values, default is 250.
 - **breaking** [all] ref: Expose `module` in `package.json` as entry point for esm builds.
+- [hub] feat: Introduce `setExtra`, `setTags`, `clearBreadcrumbs` additionally some `set` on the Scope now accept no
+  argument which makes it possible to unset the value.
 
 ## 4.6.4